Update models, views, serializers: login, create
This commit modifies and improves the models for User, Post, and Folder
and creates the Django migration necessary. It also sets up relations
for the DB schema, and creates the serializers for the models. The
views are also updated to reflect the changes in the models and
serializers.

User login and registration should now work, but the application control
flow will not be fully functional until other supporting endpoints
are completed.