Add GitHub Actions / CI & Unit Tests for Django
This change introduces GitHub Actions CI and basic unit tests for the
PosteAPI backend Django application.  It adds unit tests to test
the models and serializers, and a basic test to ensure that the
default Django project is working and models exist in the database.
Requirements have been updated, and an incorrect model was fixed.