Make user creation flexible with name
Currently, the app passes 'name' when creating a user. Ideally, we want
the android app to split the name into 'first_name' and 'last_name'
which is what Django expects. However, if we require these now, the
current application won't work.

This commit makes the user creation flexible.

If a user passes 'name' and does not provide both 'first_name' and
'last_name' (as Poste currently does) it will be split into 'first_name'
and 'last_name' by Django, and the passed 'first_name' and 'last_name'
will be ignored -- they are not required.

If a user passes 'first_name' and 'last_name' they will be used instead,
and 'name' ignored.

Either 'name' or 'first_name' and 'last_name' must be passed, or a
ValidationError will be raised. This means that there is flexibility in
what is passed.