Add Dockerization
This commit adds a Dockerfile and a docker-compose to run the application
via Docker. This stands up 3 services:

- poste: The PosteBackend Django application, served by gunicorn
- postgres: Postgres database (replaces db.sqlite3)
- nginx: Serves static and media files, acts as reverse proxy, and serves as load balancer

Currently, the Poste app sends requests to localhost:8000 -- in the future, this will be replaced
by requests to port 80 (HTTP) or 443 (HTTPS). For now, I've included a port redirect from 8000:80,
so that requests from the application are still forwarded. Once this is merged in, we should alter
the app to send requests to port 80 (and when HTTPS is implemented via certs, 443) rather than
8000, and then that line should be removed from Nginx.

Dockerfile: Builds the docker image for PosteBackend
Makefile: Convenience commands
deploy/entrypoint.sh: Entrypoint script for the PosteBackend app
deploy/nginx.conf: Nginx configuration
docker-compose.yml: Stands up PosteBackend container alongside Postgres and Nginx containers

PLEASE: See the README with any questions. The way this app works has changed and it may perform
unexpectedly.