Add tags when creating a Post
This commit allows the creation and assignment of Tags to Posts during
Post creation API call. Tags are sent as one long string, and are delineated
by commas. For example:

4 tags: "This, is, four, tags"
1 tag:  "One"
1 tag:  "This is also one tag"          (delineated by commas)
4 tags: "This, , is, , four, , tags, "  (empty entries ignored)
0 tags: ""                              (empty entries ignored)

A new Tag model was created and the Post model was modified; Tags
are unique by name, so only one tag can exist with a given name. When
assigning Tags to Posts, it will either use an existing Tag with the same
name, or create it if it does not exist. This creates a new migration,
so make sure you run `python manage.py migrate` to modify your DB schema
appropriately.

Similar logic can be used in the editing of an existing post, but is not
implemented here.