Debug share folder granting erroneous access
Previously, attempting to 'share' a folder to someone and giving them
no access actually created a folder permission; this ensures that if
we attempt to strip access from a user that any pertinent permission
is deleted.

The intent is that we should not have 'null' permissions
anywhere, and there's currently a bug where folders with null permissions
are being returned to a user during update, which is not the intended
behavior. While that bug should also be squashed, this mitigates the
problem while resolving its own issues.